ai_agent_service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-experimental
langchain-core
langchain-ollama
pydantic
python-dotenv
pymysql
sqlalchemy
requests
mysqlclient
mysql-connector-python
